Cam/pascal case (#4)
This PR adds a new PascalCase rule.
This rule errors when a variable, containing a function, doesn't have a
PascalCase name.
It's written to replace @typescript-eslint/naming-convention for our
needs, since it was taking up about 25 seconds, or about 40%, of the
time taken to lint our backend repo.

I've written documentation (maybe too much😅) in the code, so check that
our for more details